# refactor: micro optimize broad graph updates and node creation