--- a/bench/src/frameworks/zedux.ts
+++ b/bench/src/frameworks/zedux.ts
@@ -10,20 +10,29 @@ import { Computed, ReactiveFramework } from '../util/reactiveFramework'
 let ecosystem: Ecosystem
 let counter = 0
 
-class CustomSelector extends SelectorInstance {
+class BenchmarkableSelector extends SelectorInstance {
   /**
-   * Zedux cleans up graph nodes when it detects they're no longer in use. These
-   * benchmarks don't expect that behavior. Making `node.m`aybeDestroy a noop is
-   * all that's needed to prevent automatic destruction, while still allowing
-   * manual destruction via `node.destroy()`.
+   * Zedux is a cache manager. Unlike other signals-libs, it cleans up graph
+   * nodes when it detects they're no longer in use. These benchmarks don't
+   * expect that behavior. Making `node.m`aybeDestroy a noop is all that's
+   * needed to prevent automatic destruction, while still allowing manual
+   * destruction via `node.destroy()`.
    */
+  m() {}
+}
+
+class BenchmarkableSignal extends Signal {
   m() {}
 }
 
 export const zeduxFramework: ReactiveFramework = {
   name: 'Zedux',
   signal: initialValue => {
-    const s = new Signal(ecosystem, (counter++).toString(), initialValue)
+    const s = new BenchmarkableSignal(
+      ecosystem,
+      (counter++).toString(),
+      initialValue
+    )
 
     return {
       read: () => s.get(),
@@ -31,13 +40,19 @@ export const zeduxFramework: ReactiveFramework = {
     }
   },
   computed: <T>(fn: () => T): Computed<T> => {
-    const s = new CustomSelector(ecosystem, (counter++).toString(), fn, [])
+    const s = new BenchmarkableSelector(
+      ecosystem,
+      (counter++).toString(),
+      fn,
+      []
+    )
 
     return {
       read: () => s.get(),
     }
   },
-  effect: fn => new CustomSelector(ecosystem, (counter++).toString(), fn, []),
+  effect: fn =>
+    new BenchmarkableSelector(ecosystem, (counter++).toString(), fn, []),
   withBatch: fn => ecosystem.batch(fn),
   withBuild: fn => {
     if (ecosystem) ecosystem.reset()