@affects atoms, core, stores

## Description

At 1400+ ms, Zedux's `broadPropagation` benchmark score, while >4x faster than Jotai, is still far behind most other signals libraries. This is because Zedux uses a scheduler for everything. At best, this is an O(n log n) algorithm. Most other signals libs use an O(n) (2n to be precise) push-pull algorithm.

The "phase 2" of the graph algorithm change I've been talking about involves switching us to a similar model for normal graph updates. But rather than bail immediately, we should optimize this algorithm first.

The difference between O(n) and O(n log n) shouldn't be as extreme as it looks right now. In fact, we would be able to micro-optimize Zedux so much that the difference is imperceptible. However, our scheduler algorithm must use a dense JS array. And v8 doesn't provide a way to teach the engine how to optimally resize that array.

This means our `splice` and `push` operations are worse than they algorithmically need to be. The good news is that's the only downside. Other than array resizing, we can micro optimize Zedux so much that the difference between O(n) and our O(n log n) is very small.

So do that. Use an offset counter instead of shifting the jobs array when running jobs to cut the amount of array resizing in half. Replace the `findIndex` closure/recursion with a micro optimized `while` loop. Remove several intermediary variables and use bitwise operations for rounding.

Also micro optimize the `createDataSignals` benchmark further. Lazily initialize the GraphNode `o`bservers and `s`ources properties depending on the node type. For example, signals rarely have sources, while external nodes rarely have observers. Don't lazily initialize either map for selectors since they'll rarely benefit from that.

### Bottom Line

These changes improve these benchmarks:

- `avoidablePropagation` from 228 ms to 179 ms
- `broadPropagation` from 1400 ms to 742 ms. For comparison, SolidJS is at 643 ms. We're now where we should be.
- `deepPropagation` from 299 ms to 219 ms
- `diamond` from 505 ms to 385 ms
- `mux` from 434 ms to 291 ms
- `triangle` from 163 ms to 138 ms
- `createDataSignals` from ~28 ms to ~20 ms

All other benchmarks are marginally better.

<details>
  <summary>Full Kairo Bench Comparison</summary>
  <pre>framework        , test                                                         , time     , gcTime
@angular/signals , avoidablePropagation                                         , 640.54   , 121.76
@angular/signals , broadPropagation                                             , 944.50   , 24.01 
@angular/signals , deepPropagation                                              , 282.20   , 8.13  
@angular/signals , diamond                                                      , 483.01   , 56.56 
@angular/signals , mux                                                          , 823.98   , 16.03 
@angular/signals , repeatedObservers                                            , 38.61    , 5.87  
@angular/signals , triangle                                                     , 180.10   , 10.43 
@angular/signals , unstable                                                     , 62.52    , 7.36  
Jotai            , avoidablePropagation                                         , 3142.20  , 563.09
Jotai            , broadPropagation                                             , 6280.17  , 731.89
Jotai            , deepPropagation                                              , 2272.49  , 262.63
Jotai            , diamond                                                      , 3465.44  , 512.21
Jotai            , mux                                                          , 2513.68  , 200.57
Jotai            , repeatedObservers                                            , 408.73   , 56.90 
Jotai            , unstable                                                     , 821.12   , 85.63 
MobX             , avoidablePropagation                                         , 623.04   , 20.07 
MobX             , broadPropagation                                             , 1522.91  , 49.09 
MobX             , deepPropagation                                              , 594.80   , 15.82 
MobX             , diamond                                                      , 781.18   , 20.74 
MobX             , mux                                                          , 688.26   , 14.48 
MobX             , repeatedObservers                                            , 56.85    , 1.74  
MobX             , triangle                                                     , 262.09   , 6.78  
MobX             , unstable                                                     , 109.23   , 2.28  
$mol_wire_atom   , avoidablePropagation                                         , 288.26   , 8.50  
$mol_wire_atom   , broadPropagation                                             , 851.97   , 19.57 
$mol_wire_atom   , deepPropagation                                              , 212.10   , 2.58  
$mol_wire_atom   , diamond                                                      , 368.64   , 10.47 
$mol_wire_atom   , mux                                                          , 619.45   , 13.67 
$mol_wire_atom   , repeatedObservers                                            , 102.80   , 2.31  
$mol_wire_atom   , triangle                                                     , 115.14   , 1.83  
$mol_wire_atom   , unstable                                                     , 181.30   , 2.91  
Preact Signals   , avoidablePropagation                                         , 135.82   , 0.69  
Preact Signals   , broadPropagation                                             , 229.31   , 0.00  
Preact Signals   , deepPropagation                                              , 103.11   , 0.00  
Preact Signals   , diamond                                                      , 170.38   , 0.77  
Preact Signals   , mux                                                          , 241.01   , 1.67  
Preact Signals   , repeatedObservers                                            , 30.87    , 0.00  
Preact Signals   , triangle                                                     , 60.67    , 0.20  
Preact Signals   , unstable                                                     , 50.99    , 0.00  
@reactively      , avoidablePropagation                                         , 186.15   , 1.55  
@reactively      , broadPropagation                                             , 293.15   , 0.77  
@reactively      , deepPropagation                                              , 134.48   , 0.00  
@reactively      , diamond                                                      , 185.86   , 1.18  
@reactively      , mux                                                          , 263.29   , 2.00  
@reactively      , repeatedObservers                                            , 48.45    , 0.17  
@reactively      , triangle                                                     , 61.25    , 0.47  
@reactively      , unstable                                                     , 127.62   , 6.31  
s-js             , avoidablePropagation                                         , 255.20   , 0.73  
s-js             , broadPropagation                                             , 218.72   , 0.00  
s-js             , deepPropagation                                              , 68.07    , 0.00  
s-js             , diamond                                                      , 173.15   , 0.73  
s-js             , mux                                                          , 296.73   , 2.07  
s-js             , repeatedObservers                                            , 83.33    , 0.00  
s-js             , triangle                                                     , 61.31    , 0.18  
s-js             , unstable                                                     , 118.78   , 0.00  
SolidJS          , avoidablePropagation                                         , 294.30   , 4.33  
SolidJS          , broadPropagation                                             , 634.77   , 4.15  
SolidJS          , deepPropagation                                              , 249.49   , 1.92  
SolidJS          , diamond                                                      , 373.04   , 4.65  
SolidJS          , mux                                                          , 427.31   , 3.78  
SolidJS          , repeatedObservers                                            , 123.96   , 1.30  
SolidJS          , triangle                                                     , 145.96   , 2.07  
SolidJS          , unstable                                                     , 169.66   , 1.70  
Zedux            , avoidablePropagation                                         , 177.73   , 6.96  
Zedux            , broadPropagation                                             , 748.41   , 12.38 
Zedux            , deepPropagation                                              , 220.65   , 4.43  
Zedux            , diamond                                                      , 379.48   , 9.15  
Zedux            , mux                                                          , 301.38   , 5.41  
Zedux            , repeatedObservers                                            , 38.64    , 0.50  
Zedux            , triangle                                                     , 137.44   , 2.66  
Zedux            , unstable                                                     , 91.89    , 3.29  </pre>
</details>